+++ src/app/middleware/middleware.js
@@ -122,7 +122,7 @@ class Middleware {
     }
 
     // COURSE TRANSACTION
-    courseTransaction(req, res, next) {
+    CourseTransaction(req, res, next) {
         if(req.headers.authorization && req.headers.authorization !== "Empty") {
             if(JWT.verify(req.headers.authorization.split(' ')[1])) {
                 let { token, role } = JWT.decoded(req.headers.authorization.split(' ')[1]).payload;
@@ -199,22 +199,37 @@ class Middleware {
         }
     }
 
-    unitTransactions(req, res, next) {
+    UnitTransactions(req, res, next) {
         if(req.headers.authorization && req.headers.authorization !== "Empty") {
             if(JWT.verify(req.headers.authorization.split(' ')[1])) {
                 let { token, role } = JWT.decoded(req.headers.authorization.split(' ')[1]).payload;
                 req.condition = {"_id": {"$eq": new ObjectId(token)}};
 
-                if(req.headers.comment) {
-                    let {type, token, id, limited, start} = JSON.parse(req.headers.comment);
-                    req.type = "limited";
-                    req.limited = limited;
-                    req.start = start;
-                    req.courseCondition = {"CourseId": {"$eq": id}};
-                    console.log( req.courseCondition);
+                if(role) {
+                    if(role === "amin") {
+                        console.log(role);
+                        console.log(body);
+
+                    } else {
+                        if(req.headers.comment) {
+                            let {type, token, id, limited, start} = JSON.parse(req.headers.comment);
+
+                            if(type === "Find") {
+                                req.type = "Find";
+                                req.condition = {"CourseId": req.id};
+                            }
+                            next();
+        
+                        }
+                    }
+
+                } else {
+                    console.log("Not Find Role");
 
                 }
-                next();
+
+                
+                // next();
 
             } else {
                 return res.status(404).json({status: false, type: "tokenExperience"});
