+++ src/assets/js/unit/unit.js
@@ -1,6 +1,6 @@
 import { Cookie } from "../lib/cookie.js";
 import { environment } from "../config/environment.js";
-import { HTTPS } from "../commons/httpService.js";
+import { HTTPS } from "../lib/https.js";
 import { View } from "../lib/view.js";
 import Origin from "../lib/lib-origin.js";
 
@@ -8,30 +8,48 @@ window.onload = function (e) {
     const cookie = new Cookie();
     const https = new HTTPS();
     const origin = new Origin();
-    const view = new View();
+    // const view = new View();
     const $ = document.querySelector.bind(document);
     const $$ = document.querySelectorAll.bind(document);
 
-        let token = `Bearer ${cookie.get("Authentic")}`;
-        environment.payload.type = "limited";
-        environment.payload.start = 0;
-        environment.payload.limited = 5;
-        environment.payload.id = origin.parameter().token;
+    let token = `Bearer ${cookie.get("Authentic")}`;
 
-        let ComponentHeader = $("#Header");
-        let ComponentView = $("#Body");
-        let KeyComponent = ["Name", "AmountLesson", "Status", "CreateDate"];
-        let KeyHeader = ["Tên học phần", "Số khóa học", "Trạng thái", "Ngày tạo", "Chúc năng"];
-
-        (function() {
-            https.FIND(environment.payload, token, environment.endpoint.unit)
-            .then((res) => {
-                view.render(res, ComponentHeader, KeyHeader, ComponentView, KeyComponent);
+    (function(){
+        if(localStorage.getItem("UnitToken")) {
+            let payload = {
+                id: localStorage.getItem("UnitToken"),
+                type: "Find"
+            }
+            https.FIND(payload, token, environment.endpoint.unit)
+            .then((result) => {
+                console.log(result);
             })
             .catch((err) => {
                 throw err;
             })
-        }());
+        }
+    }())
+
+    //     let token = `Bearer ${cookie.get("Authentic")}`;
+    //     environment.payload.type = "limited";
+    //     environment.payload.start = 0;
+    //     environment.payload.limited = 5;
+    //     environment.payload.id = origin.parameter().token;
+
+    //     let ComponentHeader = $("#Header");
+    //     let ComponentView = $("#Body");
+    //     let KeyComponent = ["Name", "AmountLesson", "Status", "CreateDate"];
+    //     let KeyHeader = ["Tên học phần", "Số khóa học", "Trạng thái", "Ngày tạo", "Chúc năng"];
+
+    //     (function() {
+    //         https.FIND(environment.payload, token, environment.endpoint.unit)
+    //         .then((res) => {
+    //             view.render(res, ComponentHeader, KeyHeader, ComponentView, KeyComponent);
+    //         })
+    //         .catch((err) => {
+    //             throw err;
+    //         })
+    //     }());
 
 //     (function () {
 //         routerNew.setAttribute("href", location.href.replace("?", "/detail?type=create&"));
