+++ src/assets/js/unit/unitDetail.js
@@ -2,48 +2,45 @@
 // import { httpsService } from "../commons/httpService.js";
 // import { Validation } from "../commons/validation.js";
 
-// window.onload = function (e) {
+// window.onload = function () {
 //     const $ = document.querySelector.bind(document);
 //     const $$ = document.querySelectorAll.bind(document);
 
 //     // GET WRAPPER
-//     let lessonForm = $("#lesson-detail");
-//     let unitCode = $("#unit-code");
-//     let name = $("#lesson-name");
-//     let content = $("#lesson-content");
-//     let status = $("input[name='status']:checked");
-//     let thumbnail = $("#lesson-thumbanil");
-//     let createDate = $("#lesson-create-date");
-//     let editLastDate = $("#lesson-edit-last-date");
+//     let unitForm = $("#unit-detail");
+//     let courseCode = $("#course-code");
+//     let unitName = $("#unit-name");
+//     let unitLesson = $("#unit-lesson");
+//     let unitStatus = $("input[name='status']:checked");
+//     let unitCreateDate = $("#unit-create-date");
+//     let unitEditLastDate = $("#unit-edit-last-date");
 
 //     let toasts = $$(".modal-toasts")[0];
 
-//     $("#go-back").addEventListener("click", function (element) {
-//         window.history.back();
-//     })
+//     $$(".go-back").forEach((e) => {
+//         e.addEventListener("click", function (element) {
+//             window.history.back();
+//         })
+//     });
 
 //     Validation({
-//         form: "#lesson-detail",
+//         form: "#unit-detail",
 //         selectorError: ".form-message",
 //         rules: [
 //             {
-//                 selector: "#lesson-name",
+//                 selector: "#course-code",
 //                 guides: [Validation.required()]
 //             },
 //             {
-//                 selector: "#lesson-content",
+//                 selector: "#unit-name",
 //                 guides: [Validation.required()]
 //             },
 //             {
-//                 selector: "#lesson-thumbanil",
+//                 selector: "#unit-create-date",
 //                 guides: [Validation.required()]
 //             },
 //             {
-//                 selector: "#lesson-create-date",
-//                 guides: [Validation.required()]
-//             },
-//             {
-//                 selector: "#lesson-edit-last-date",
+//                 selector: "#unit-edit-last-date",
 //                 guides: [Validation.required()]
 //             }
 //         ]
@@ -52,7 +49,7 @@
 //     if (getType() == "update") {
 //         (function () {
 //             console.log(getToken());
-//             httpsService("API/lesson/lesson-single", "POST", { id: getToken() })
+//             httpsService("API/unit/unit-single", "POST", { id: getToken() })
 //                 .then((data) => {
 //                     return data.json();
 //                 })
@@ -68,29 +65,29 @@
 //     switch (getType()) {
 //         case "update":
 //             setTitleForm("update");
-//             lessonForm.addEventListener("submit", updateLesson);
+//             unitForm.addEventListener("submit", updateUser);
 //             break;
 
 //         case "create":
 //         default:
-//             unitCode.value = getToken();
+//             courseCode.value = getToken();
 //             setTitleForm("create");
-//             lessonForm.addEventListener("submit", createLesson);
+//             unitForm.addEventListener("submit", createUser);
 //             break;
 //     }
 
-//     function createLesson(e) {
+//     function createUser(e) {
 //         e.preventDefault();
 //         if (this.valid) {
-//             let lesson = getLesson();
-//             if (lesson) {
-//                 httpsService("API/lesson/lesson-new", "POST", lesson)
+//             let unit = getCourseForm();
+//             if (unit) {
+//                 httpsService("API/unit/unit-new", "POST", unit)
 //                     .then((res) => {
 //                         return res.json();
 //                     })
 //                     .then((data) => {
 //                         data.status ?
-//                             location.href = `/courses/units/lessons?unitId=${lesson.unitId}` :
+//                             location.href = `/courses/units?course=${unit.courseId}` :
 //                             permission(toasts, data);
 //                     })
 //                     .catch((err) => {
@@ -100,19 +97,19 @@
 //         }
 //     }
 
-//     function updateLesson(e) {
+//     function updateUser(e) {
 //         e.preventDefault();
 //         if (this.valid) {
-//             let lesson = getLesson();
-//             lesson["id"] = getToken();
-//             if (lesson) {
-//                 httpsService("API/lesson/lesson-edit", "PUT", lesson)
+//             let unit = getCourseForm();
+//             unit["id"] = getToken();
+//             if (unit) {
+//                 httpsService("API/unit/unit-edit", "PUT", unit)
 //                     .then((res) => {
 //                         return res.json();
 //                     })
 //                     .then((data) => {
 //                         data.status ?
-//                             location.href = `/courses/units/lessons?unitId=${lesson.unitId}` :
+//                             location.href = `/courses/units?course=${unit.courseId}` :
 //                             permission(toasts, data);
 //                     })
 //                     .catch((err) => {
@@ -122,27 +119,25 @@
 //         }
 //     }
 
-//     function getLesson() {
+//     function getCourseForm() {
 //         let data = {
-//             unitId: unitCode.value,
-//             lessonName: name.value,
-//             lessonContent: content.value,
+//             courseId: courseCode.value,
+//             unitName: unitName.value,
 //             status: ($("input[name='status']:checked").value == "action") ? true : false,
-//             thumbnail: thumbnail.value,
-//             createDate: createDate.value,
-//             updateDate: editLastDate.value
+//             amountLesson: (getType() == "update") ? unitLesson.value : 0,
+//             createDate: unitCreateDate.value,
+//             updateDate: unitEditLastDate.value
 //         }
 //         return data;
 //     }
 
-//     function setCourseForm(lesson) {
-//         unitCode.value = lesson.unitId;
-//         name.value = lesson.lessonName;
-//         content.value = lesson.lessonContent;
-//         (lesson.status) ? $(`input[id='active']`).checked = true : $(`input[id='no-active']`).checked = true;
-//         thumbnail.value = lesson.thumbnail;
-//         createDate.value = lesson.createDate.split(".")[0];
-//         editLastDate.value = lesson.updateDate.split(".")[0];
+//     function setCourseForm(unit) {
+//         courseCode.value = unit.courseId;
+//         unitName.value = unit.unitName;
+//         unitLesson.value = unit.amountLesson;
+//         (unit.status) ? $(`input[id='active']`).checked = true : $(`input[id='no-active']`).checked = true;
+//         unitCreateDate.value = unit.createDate.split(".")[0];
+//         unitEditLastDate.value = unit.updateDate.split(".")[0];
 //     }
 
 //     function setTitleForm(type) {
@@ -150,7 +145,7 @@
 //         let subButton = $$(".btn-executed")[0];
 
 //         if (type == "create") {
-//             title.innerHTML = "Thêm mới bài học";
+//             title.innerHTML = "Thêm mới Chương học";
 //             subButton.innerHTML = "Thêm mới";
 //         } else {
 //             title.innerHTML = "Chỉnh sửa thông tin";
