+++ src/assets/js/user/userDetail.js
@@ -1,10 +1,10 @@
 import { Cookie } from "../lib/cookie.js";
 import DateTimes from "../lib/date.js";
-import { HTTPS } from "../commons/https.js";
+import { HTTPS } from "../lib/https.js";
 import Origin  from "../lib/lib-origin.js";
 import { Priture } from "../lib/priture.js";
 import { environment } from "../config/environment.js";
-import { Validation } from "../commons/validation.js";
+import { Validation } from "../lib/validation.js";
 
 window.onload = function (e) {
     const cookie = new Cookie();
