+++ src/resource/views/components/course-component/unit-component/unit-component.hbs
@@ -38,7 +38,7 @@
         <ul class="pagination d-flex align-items-center justify-content-center"></ul>
     </nav>
 
-    <script type="module" src="/static/js/course/unit.js"></script>
+    <script type="module" src="/static/js/unit/unit.js"></script>
 </body>
 
 </html>
