+++ src/resource/views/components/course-component/unit-component/unit-detail-component.hbs
@@ -75,7 +75,7 @@
         </div>
     </div>
 
-    <script type="module" src="/static/js/course/unitDetail.js"></script>
+    <script type="module" src="/static/js/unit/unitDetail.js"></script>
 </body>
 
 </html>
